Refactor module structure by renaming and reorganizing imports, removing unused files, and adding new utility functions for improved clarity and functionality
