+++ ocean/__init__.py
@@ -1,10 +1,10 @@
-from . import datasets, ensemble, feature, mip, tree
+from . import abc, datasets, feature, mip, tree
 from .explainer import MixedIntegerProgramExplainer
 
 __all__ = [
     "MixedIntegerProgramExplainer",
+    "abc",
     "datasets",
-    "ensemble",
     "feature",
     "mip",
     "tree",
