--- ocean/ensemble/__init__.py
@@ -1,3 +1,0 @@
-from .ensemble import Ensemble
-
-__all__ = ["Ensemble"]
