+++ ocean/mip/_typing.py
@@ -1,0 +1,3 @@
+import gurobipy as gp
+
+type Objective = gp.LinExpr | gp.QuadExpr
