+++ ocean/mip/_variable/__init__.py
@@ -1,0 +1,4 @@
+from ._feature import FeatureVar
+from ._tree import TreeVar
+
+__all__ = ["FeatureVar", "TreeVar"]
