--- ocean/mip/variable/__init__.py
@@ -1,4 +1,0 @@
-from .feature import FeatureVar
-from .tree import TreeVar
-
-__all__ = ["FeatureVar", "TreeVar"]
