+++ ocean/tree/__init__.py
@@ -1,5 +1,5 @@
-from .node import Node
-from .parse import parse_tree, parse_trees
-from .tree import Tree
+from ._node import Node
+from ._parse import parse_ensembles, parse_tree, parse_trees
+from ._tree import Tree
 
-__all__ = ["Node", "Tree", "parse_tree", "parse_trees"]
+__all__ = ["Node", "Tree", "parse_ensembles", "parse_tree", "parse_trees"]
