+++ ocean/tree/_node.pyi
