+++ tests/mip/model/test_feasible.py
@@ -1,13 +1,9 @@
-from functools import partial
-from itertools import chain
-
 import gurobipy as gp
 import numpy as np
 import pytest
 
-from ocean.ensemble import Ensemble
 from ocean.mip import Model
-from ocean.tree import parse_trees
+from ocean.tree import parse_ensembles, parse_trees
 
 from ...utils import ENV
 from ..utils import (
@@ -142,8 +138,7 @@ class TestIsolation:
             n_samples,
             n_classes,
         )
-        parser = partial(Ensemble, mapper=mapper)
-        trees = chain.from_iterable(map(parser, (clf, ilf)))
+        trees = parse_ensembles(clf, ilf, mapper=mapper)
         model = Model(
             trees=trees,
             mapper=mapper,
@@ -185,8 +180,7 @@ class TestIsolation:
             n_classes,
             return_data=True,
         )
-        parser = partial(Ensemble, mapper=mapper)
-        trees = chain.from_iterable(map(parser, (clf, ilf)))
+        trees = parse_ensembles(clf, ilf, mapper=mapper)
         model = Model(
             trees=trees,
             mapper=mapper,
